suggest libreoffice GTK2 / remove KMFL restart
The GTK3 theme is terrible for wasta users. It doesn't fit well
with the Cinnamon / Windows look and feel, so tell LO to use
the GTK2 look instead - which is what Nemo is based on. Doing
it in /etc/profile.d is great because it isn't affected by
updates, and even "fixes" my compiled development starts.

KMFL needs to be restarted, but a killall -SIGHUP doesn't seem
to do the job, and just doing an ibus-daemon -xrd runs only
as root (the only user dpkg knows about). So just remove it
for now unless someone has a great idea on how to force
ibus-daemon to restart for all logged in users without starting
a separate instance for root.