general cleanup and tweaking: remove old LO 5.4 code
1.) LibreOffice 5.4 I think was the base for 16.04.
In any case, 6.1 and 6.2 are available for it,
so very little reason to worry about 5.4 anymore,
since this template is only going to apply to 2020
situations where xenial is nearly obsolete.

2.) Add explanation why it should be a default suggestion
to disable gnome-software doing its own updates.
Now if only I could figure out why ubuntu-software is also
doing updates - and how to disable that.

3.) lintian warning: shlibs is not a depends for shell scripts?

4.) change -template to -BRANCH_ID to limit the need to
search/replace multiple strings.