polkit .rules (for 24.04+) since .pkla files are now obsolete
AFAIK, .pkla is ignored after .105,
and .rules were ignore by .105.
Ubuntu/Debian were stuck on .105 for a very long time,
but now polkit has moved fully to javascript definitions.

Search for "polkit_pkla2rule.py" as a possible converter
(although it didn't handle unix-group:*).

Permissions are important. I think the most proper is
440 root:polkitd, but a 444 should work fine.

systemctl restart polkit is very important (despite
/var/logs/auth.log indicating that the rules change on every save).

It was hard to debug the rules. I tried lots of web suggestions,
but in the end, I killed polkitd and ran it interactively.
(I don't know if the other stuff I did actually came into play...)

G_MESSAGES_DEBUG=all /usr/lib/policykit-1/polkitd

A third annoying thing is that it seems to be necessary to enclose
in round brackets the == clauses, so I did that already
inside the if() clause to encourage keeping it in case a
&& subject.isInGroup("sudo")
or some other condition is added.